Add script to download integration test binaries for next-gen (#2770)
close pingcap/ticdc#2771

+++ tests/integration_tests/debezium02/run.sh
@@ -12,6 +12,10 @@ if [ "$SINK_TYPE" != "kafka" ]; then
 	exit 0
 fi
 
+if [ "$NEXT_GEN" = 1 ]; then
+	exit 0
+fi
+
 stop_tidb_cluster
 rm -rf $WORK_DIR && mkdir -p $WORK_DIR
